Expense CRUD Done

+++ webapp/src/app/car_showroom_accounting_system/views/expense/expense-list/expense-list.component.ts
@@ -3,6 +3,7 @@ import { Router } from '@angular/router';
 import { ToastrService } from 'ngx-toastr';
 import { ExpenseService } from 'src/app/car_showroom_accounting_system/Services/expense.service';
 import {Location} from '@angular/common';
+import Swal from 'sweetalert2';
 
 @Component({
   selector: 'app-expense-list',
@@ -41,16 +42,28 @@ export class ExpenseListComponent implements OnInit {
 
   deleteExpense(link: string)
   {
+    Swal.fire({
+    title: 'Are you sure?',
+    text: 'You won\'t be able to revert this!',
+    icon: 'warning',
+    showCancelButton: true,
+    confirmButtonColor: '#d33',
+    cancelButtonColor: '#B3BBC2',
+    confirmButtonText: 'Yes, delete it!'
+  }).then((result) => {
+    if (result.isConfirmed) {
     this.expenseService.deleteExpense(link)
     .subscribe(
       data => {
-        this.toastrService.success('Purchase Record Delete Successful.', 'Success');
+        this.toastrService.success('Expense Record Delete Successful.', 'Success');
         this.ngOnInit();
       },
       error => {
         this.toastrService.error(error.error);
       }
     );
+    }
+  });
   }
 
   expenseDetails(link: string)
